BugFix: NextpeerNotifier queueScheduler should accept no param; Add Nextpeer getClass JNI as Cocos2d-x does not support that any more

+++ Nextpeer-Cocos2D-X/Classes/NextpeerEventQueue.cpp
@@ -32,7 +32,7 @@ namespace nextpeer {
         pthread_mutex_unlock(&_lock);
     }
     
-    unsigned int NextpeerEventQueue::queueSize()
+    unsigned long NextpeerEventQueue::queueSize()
     {
         return _queue.size();
     }
